Fix undeclared dependency on mongodb module
The mongodb module is imported from src/cache.js but isn't listed as a dependeny in package.json.

Instead of hauling in the mongodb module just for its ObjectId implementation, use ObjectId from the bson module, which is already a dependency.

This requires an implementation of isObjectId(any) => boolean (with duck typing) to use instead of instanceof ObjectId, as ObjectId instances from the bson module won't match instances from other modules. isObjectId() is added to src/helpers.js.